devices: dku2libusb: move definitions into C file
